Version 1.300
General
	• Config-file is restructured and expanded to take more than one
parameter set and quick test templates.
	• When opening new images the append/replace question is no longer
displayed. Instead an append-button can be set and saved as default.
	• Export settings is taken from the Settings menu onto the main window
for easier access.
	• Added option to sort images by filename, acquisition time, series
number, image number or z position.
	• The mode changes now automatically  according to the modality of the
images (DR,CR,CT,NM,PT) when opening images. If append is used, this is
ignored. The user may also change mode as before. 
	• SPECT is now regarded as a separate modality in ImageQC by finding
modality NM together with a slicethickness for the image.
	• Set and get center buttons replaced with buttons with icon-images
(arrow and spyglass).

CT
	• Slicethickness: Added beaded ramp with two vertical ramps (for use
with GE phantom). These are drilled holes and thus added option that
ramp density is lower than background.
	• QuickTest option for CT now possible. This results in som reordering
of test as not all tests are prepared for the QuickTest option.
	• MTF: Offset for center of ROI from center of image now possible. This
was done to prepare for QuickTest using different center for MTF
evaluation than for other tests.
	
Xray
	• MTF: Offset for center of ROI from center of image now possible. This
was done to prepare for QuickTest using different center for MTF
evaluation than for other tests.

NM planar
	• SNI calculation to test for structured noise according to Nelson et
al, J Nucl Med 2014;55:169-174.
	• Uniformity-test is replaced by the NEMA calculations for uniformity.
	• Added frame duration and termination conditions to parameters read
from header. Displayed in text-info for planar NM, not for SPECT.

SPECT
	• Separated from NM. If image has modality NM and have slice thickness
the mode changes automatically to SPECT as the image is opened.