Version 1.40
General
	• Automation template for  loading and analysing images. Targeted at
frequent constancy controls with the same number of images analysed with
the same set of tests each time. Description on how we use this in
Stavanger see the Wiki.
	• Multiframe images can now be opened together with other files
(single- or multiframe).
	• Multiframe images is no longer restricted from sorting and frame
number is added as sort-option.
	• Multiframe images is no longer restricted from QuickTest and the
multimark-option.
	• Export option to transpose table - primarily targeted on the
QuickTest export choosing between row or column export. Column is
default. With many measurements row will be better to work with in
Excel.
	• Moving the application on screen is tracked so that popup-windows
follow this replacement.
	• Works with smaller screensizes (scrollbars working better).
	• Moved some options from the settings menu into the Edit/manage
settings dialogbox.
	• Different sized images can now be calulated for some tests:
homogeneity, STP, noise
	• Removed the userdefined ROI test as it is not working properly. Might
be included back and optimalized on request.
	• Some bugfixes (and probably some new and fresh bugs).

CT
	• The material and position table for the CT Number test (sensitometry
test) can now be imported from a set of Catphan versions. Changes can be
done and saved together with the parameterset (edit/manage settings).
	
NM planar
	• SNI results can now be displayed in image results as 2dNPS (filtered
by human visual response function) for the selected region (in the
table). If no region is selected in the table, the filtered 2d NPS for
the region with the maximum SNI-value is displayed in image results.
	• Added test where total count and frame duration is extracted as
table.