Version 1.89.1
Bugfix:
Introduced GE DICOM info about IR/AI strength and with that introduced
error for some older GE. Fixed.

NM - Uniformity and SNI
Option to view the curvature corrected image in the Image Result tab
Option to plot the curvature correction to verify the correction (only
in the Uniformity tab)
SNI - more options to change parameters and to visualize steps in the
calculation
SNI - estimated quantum noise now corrected by the mean value of the
correction matrix within the ROI. (Saw that this was needed when
watching the NPS curves and corresponding estimated quantum noise
portion).